Refactoring ERC165 so our proxy contract is happy
