+++ package.json
@@ -14,12 +14,11 @@
     "migrate:reset:staging": "npm run clean && npm run lint && truffle migrate --reset --network rinkeby && npm run copy:staging",
     "migrate:production": "npm run lint && truffle migrate --network mainnet && npm run copy:production",
     "migrate:reset:production": "npm run clean && npm run lint && truffle migrate --reset --network mainnet && npm run copy:production",
-    "test:title": "truffle test --network ganache ./test/ERC721/CodexTitle.test.js",
-    "test:basic": "truffle test --network ganache ./test/ERC721/ERC721BasicToken.test.js",
-    "test:proxy": "truffle test --network ganache ./test/ERC721/CodexTitleProxy.test.js",
-    "test:token": "truffle test --network ganache ./test/ERC721/ERC721Token.test.js",
-    "test:165": "truffle test --network ganache ./test/ERC165/SupportsInterface.test.js",
-    "test:token-proxy": "truffle test --network ganache ./test/ERC721/ERC721TokenProxy.test.js",
+    "test:title": "truffle test --network ganache ./test/token/CodexTitle.test.js",
+    "test:basic": "truffle test --network ganache ./test/token/ERC721BasicToken.test.js",
+    "test:proxy": "truffle test --network ganache ./test/token/CodexTitleProxy.test.js",
+    "test:token": "truffle test --network ganache ./test/token/ERC721Token.test.js",
+    "test:token-proxy": "truffle test --network ganache ./test/token/ERC721TokenProxy.test.js",
     "test": "truffle test --network ganache",
     "lint": "npm run lint:js && npm run lint:sol",
     "lint:js": "eslint .",
