+++ test/library/Pausable.test.js
