+++ test/token/ERC721TokenProxy.test.js
@@ -2,7 +2,7 @@ import assertRevert from '../helpers/assertRevert'
 import shouldBehaveLikeERC721BasicToken from './behaviors/ERC721BasicToken.behavior'
 import shouldBehaveLikeERC721Token from './behaviors/ERC721Token.behavior'
 import shouldMintERC721Token from './behaviors/ERC721Mint.behavior'
-import shouldBehaveLikeERC165 from '../ERC165/behaviors/ERC165.behavior'
+import shouldBehaveLikeERC165 from './behaviors/ERC165.behavior'
 
 const { BigNumber } = web3
 const ERC721Token = artifacts.require('ERC721TokenMock.sol')
