+++ test/token/behaviors/ERC721BasicToken.behavior.js
