+++ test/token/behaviors/ERC721Token.behavior.js
